+++ app/core/nice/event/cooltime.py
@@ -9,14 +9,12 @@ def get_nice_event_cooltime(
     region: Region,
     cooltime: MstEventCooltimeReward,
     gift_data: GiftData,
-    common_release: MstCommonRelease,
     raw_releases: list[MstCommonRelease],
 ) -> NiceEventCooltimeReward:
     return NiceEventCooltimeReward(
         spotId=cooltime.spotId,
         lv=cooltime.lv,
         name=cooltime.name,
-        commonRelease=get_nice_common_release(common_release),
         releaseConditions=[
             get_nice_common_release(release)
             for release in raw_releases